Fix hopefully the last Clang-Tidy suggestions
There were some last second things I missed.

Among these warnings were "redundant static casts" and "deprecated C-headers"

There were others too but those warnings were false flags so I kept the ones that actually made sense.

+++ include/boost/math/special_functions/fpclassify.hpp
@@ -77,7 +77,7 @@ is used.
 */
 
 #if defined(_MSC_VER) || defined(BOOST_BORLANDC)
-#include <float.h>
+#include <cfloat>
 #endif
 #ifdef BOOST_MATH_USE_FLOAT128
 #ifdef __has_include
@@ -115,7 +115,7 @@ inline bool is_nan_helper(T t, const std::true_type&)
    (void)t;
    return false;
 #else // BOOST_HAS_FPCLASSIFY
-   return (BOOST_FPCLASSIFY_PREFIX fpclassify(t) == (int)FP_NAN);
+   return (BOOST_FPCLASSIFY_PREFIX fpclassify(t) == FP_NAN);
 #endif
 }
